Fix QuantileDigestIterator bugs from #115
The QuantileDigestIterator claims to perform a post-order traversal,
but it actually has been performing in-order traversal. This is
troublesome for most Q-Digest algorithms, which depend on post-order
traversal.

The `upperBoundQuantile()` and `lowerBoundQuantile()` functions were also
not correct and have been removed and replaced with a simple
`cumulativeProportion()`. (To obtain lower bound and upper bound quantiles,
you really need to separately perform forward and reverse iteration.)

These are breaking changes, but the functions have no known usage.